+++ BOCCHI.Automator/Services/Paths/PathCalculator.cs
@@ -1,4 +1,3 @@
-using BOCCHI.Automator.Services;
 using BOCCHI.Common.Data.Aethernet;
 using BOCCHI.Common.Config;
 using BOCCHI.Common.Data.CriticalEncounters;
@@ -367,7 +366,7 @@ public class PathCalculator
         GraphTraverser traverser = new(graph, pathfinder, logger);
         traverser.AddCalculator(new WalkTeleportWalkCalculator());
         traverser.AddCalculator(new DirectWalkCalculator());
-        if (addReturnCalculator && !ReturnDelay.IsOnCooldown())
+        if (addReturnCalculator)
         {
             traverser.AddCalculator(new ReturnTeleportWalkCalculator());
         }